feat(app): rename app
