--- apps/mobile/Gemfile
@@ -1,6 +1,0 @@
-source 'https://rubygems.org'
-
-# You may use http://rbenv.org/ or https://rvm.io/ to install and use this version
-ruby File.read(File.join(__dir__, '.ruby-version')).strip
-
-gem 'cocoapods', '~> 1.11', '>= 1.11.3'
