+++ apps/mobile/android/app/build.gradle
@@ -93,9 +93,9 @@ android {
 
     compileSdkVersion rootProject.ext.compileSdkVersion
 
-    namespace "com.app"
+    namespace "me.superlight.app"
     defaultConfig {
-        applicationId "com.app"
+        applicationId "me.superlight.app"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
