+++ apps/mobile/android/settings.gradle
@@ -1,6 +1,6 @@
-rootProject.name = 'app'
+rootProject.name = 'Superlight'
 apply from: file("../node_modules/@react-native-community/cli-platform-android/native_modules.gradle"); applyNativeModulesSettingsGradle(settings)
-include ':app'
+include ':Superlight'
 includeBuild('../node_modules/react-native-gradle-plugin')
 
 apply from: new File(["node", "--print", "require.resolve('expo/package.json')"].execute(null, rootDir).text.trim(), "../scripts/autolinking.gradle")
