+++ apps/mobile/ios/Superlight.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
