+++ apps/mobile/ios/Superlight/AppDelegate.h
