+++ apps/mobile/ios/Superlight/Images.xcassets/AppIcon.appiconset/Contents.json
