+++ apps/mobile/ios/Superlight/Images.xcassets/Contents.json
