+++ apps/mobile/ios/SuperlightTests/Info.plist
