+++ apps/mobile/ios/SuperlightTests/SuperlightTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface appTests : XCTestCase
+@interface SuperlightTests : XCTestCase
 
 @end
 
-@implementation appTests
+@implementation SuperlightTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL (^)(UIView *view))test
 {
