+++ apps/mobile/src/screens/wallets/tabs/send.screen.tsx
@@ -1,4 +1,4 @@
-import { Text, View } from 'custom/styled-react-native';
+import { Text, View } from 'util/wrappers/styled-react-native';
 
 const Send = () => {
   return (
