+++ apps/mobile/src/util/navigation/wallet-navigation.ts
