+++ apps/mobile/tsconfig.json
@@ -16,7 +16,7 @@
     "paths": {
       "components/*": ["src/components/*"],
       "screens/*": ["src/screens/*"],
-      "custom/*": ["src/wrappers/*"]
+      "util/*": ["src/util/*"]
     }
   },
   "include": ["src/**/*"],
